Override $CARGO_HOME only if cargo-home-dir is in .servobuild
This preserved the fix from #11097 in the situation where `$CARGO_HOME` and
`cargo-home-dir` in `.servobuild` are both present: `.servobuild` should win.

But it changes the behavior when `$CARGO_HOME` is present and `.servobuild` is
not:  Then `$CARGO_HOME` should be used.

We now check the following values in order of priority and use the first one
that is found:

1. `cargo-home-dir` in `.servobuild`
2. `CARGO_HOME_DIR` in the enivironment
3. default value (`<servo-repo>/.cargo`)